bigdata4/requirements.txt
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asyncssh==2.21.0
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clickhouse-connect==0.8.18
colorama==0.4.6
configobj==5.0.9
coverage==7.6.12
cryptography==43.0.3
debugpy==1.6.7
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
python-dotenv==1.1.1
dpath==2.2.0
dulwich==0.23.2
dvc==3.61.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.115.9
filelock==3.18.0
flatten-dict==0.4.2
fonttools==4.56.0
fsspec==2025.5.1
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
grandalf==0.8
gto==1.7.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hvac==2.3.0
hydra-core==1.3.2
idna==3.10
iniconfig==2.1.0
ipykernel==6.26.0
ipython==8.17.0
iterative-telemetry==0.0.10
jedi==0.18.2
joblib==1.4.2
jupyter_client==8.3.0
jupyter_core==5.3.0
kafka-python==2.2.0
kiwisolver==1.4.8
kombu==5.5.4
lz4==4.4.4
markdown-it-py==3.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mock==5.1.0
multidict==6.6.3
nest_asyncio==1.5.8
networkx==3.5
numpy==2.2.3
oauth2client==4.1.3
omegaconf==2.3.0
orjson==3.10.18
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
pluggy==1.6.0
prompt_toolkit>=3.0.38,<3.1
protobuf==6.31.1
psutil==5.9.6
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.1
PyDrive2==1.21.3
pygit2==1.18.0
Pygments==2.16.1
pygtrie==2.5.0
pyOpenSSL==24.2.1
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dateutil==2.8.2
pytz==2025.1
requests==2.32.4
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shortuuid==1.0.13
six==1.16.0
sqltrie==0.11.2
starlette==0.45.3
tabulate==0.9.0
threadpoolctl==3.5.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing_extensions>=4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.8
yarl==1.20.1
zipp==3.18.0
zstandard==0.23.0
